# packages/contracts/openapi/admin-api.yaml
openapi: "3.1.0"
info:
  title: Gantry Admin API
  version: "1.0.0"
  description: Administrative API for immutable agent lifecycle management.
servers:
  - url: /api/admin/v1
security:
  - oauth2: [admin]
paths:
  /workspaces:
    get:
      operationId: listManagedWorkspaces
      summary: List workspaces manageable by the current administrator
      tags: [Workspaces]
      responses:
        "200":
          description: Managed workspaces
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WorkspaceList"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /agents:
    get:
      operationId: listAgents
      summary: List agents in manageable workspaces
      tags: [Agents]
      parameters:
        - name: workspace_id
          in: query
          schema: {type: string}
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AgentList"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
    post:
      operationId: createAgent
      summary: Create an agent and its initial draft
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateAgentRequest"}
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Agent"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/InvalidInput"}
  /agents/{agent_id}:
    get:
      operationId: getAgent
      summary: Get an agent in the current administrator domain
      tags: [Agents]
      parameters: [{$ref: "#/components/parameters/AgentId"}]
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Agent"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /agents/{agent_id}/draft:
    get:
      operationId: getAgentDraft
      summary: Get the editable current draft
      tags: [Drafts]
      parameters: [{$ref: "#/components/parameters/AgentId"}]
      responses:
        "200":
          description: Draft
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AgentDraft"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
    put:
      operationId: updateAgentDraft
      summary: Replace an agent draft with optimistic concurrency
      tags: [Drafts]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateDraftRequest"}
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AgentDraft"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "412": {$ref: "#/components/responses/RevisionConflict"}
        "422": {$ref: "#/components/responses/InvalidInput"}
  /agents/{agent_id}/versions:
    get:
      operationId: listAgentVersions
      summary: List immutable versions
      tags: [Versions]
      parameters: [{$ref: "#/components/parameters/AgentId"}]
      responses:
        "200":
          description: Version history
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AgentVersionList"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /agents/{agent_id}/review:
    get:
      operationId: getAgentReview
      summary: Get the current draft review and semantic diff
      tags: [Reviews]
      parameters: [{$ref: "#/components/parameters/AgentId"}]
      responses:
        "200":
          description: Current review state
          content: {application/json: {schema: {$ref: "#/components/schemas/AgentReview"}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /agents/{agent_id}:publish:
    post:
      operationId: publishAgentDraft
      summary: Freeze and publish the current valid draft
      tags: [Publications]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Version from an idempotent publication retry
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AgentVersion"}
        "201":
          description: Published immutable version
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AgentVersion"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/InvalidState"}
        "412": {$ref: "#/components/responses/RevisionConflict"}
  /agents/{agent_id}:review:
    post:
      operationId: submitAgentReview
      summary: Submit the current draft for review
      tags: [Reviews]
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/SubmitReviewRequest"}}}
      responses:
        "201":
          description: Review submitted
          content: {application/json: {schema: {$ref: "#/components/schemas/AgentReview"}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/InvalidState"}
        "412": {$ref: "#/components/responses/RevisionConflict"}
  /agents/{agent_id}:review-decision:
    post:
      operationId: decideAgentReview
      summary: Approve or reject the current draft review
      tags: [Reviews]
      parameters: [{$ref: "#/components/parameters/AgentId"}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/ReviewDecisionRequest"}}}
      responses:
        "200":
          description: Review decision recorded
          content: {application/json: {schema: {$ref: "#/components/schemas/AgentReview"}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/InvalidState"}
  /agents/{agent_id}:retire:
    post:
      operationId: retireAgent
      summary: Retire the currently published agent
      tags: [Publications]
      parameters: [{$ref: "#/components/parameters/AgentId"}]
      responses:
        "204": {description: Agent retired}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/InvalidState"}
  /agents/{agent_id}:rollback:
    post:
      operationId: rollbackAgent
      summary: Roll back to an immutable published version
      tags: [Publications]
      parameters: [{$ref: "#/components/parameters/AgentId"}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/RollbackRequest"}}}
      responses:
        "204": {description: Agent rolled back}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/InvalidState"}
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes: {admin: Administrative agent lifecycle access}
  parameters:
    AgentId:
      name: agent_id
      in: path
      required: true
      schema: {type: string}
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string}
      description: Current draft revision.
  responses:
    Unauthorized:
      description: Authentication is required.
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
    Forbidden:
      description: Administrative access is required.
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
    NotFound:
      description: Resource was not found.
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
    InvalidInput:
      description: Request is invalid.
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
    InvalidState:
      description: Operation is not permitted in the current state.
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
    RevisionConflict:
      description: Draft revision is stale.
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
    PageInfo:
      type: object
      required: [has_more]
      properties:
        has_more: {type: boolean}
    Workspace:
      type: object
      required: [id, slug, display_name]
      properties:
        id: {type: string}
        slug: {type: string}
        display_name: {type: string}
    WorkspaceList:
      type: object
      required: [items, page_info]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/Workspace"}}
        page_info: {$ref: "#/components/schemas/PageInfo"}
    Agent:
      type: object
      required: [id, organization_id, workspace_id, slug, display_name, description, category, lifecycle_status]
      properties:
        id: {type: string}
        organization_id: {type: string}
        workspace_id: {type: string}
        slug: {type: string}
        display_name: {type: string}
        description: {type: string}
        category: {type: string}
        lifecycle_status: {type: string, enum: [draft, published, retired]}
        current_published_version_id: {type: string}
    AgentList:
      type: object
      required: [items, page_info]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/Agent"}}
        page_info: {$ref: "#/components/schemas/PageInfo"}
    CreateAgentRequest:
      type: object
      required: [workspace_id, slug, display_name, description, category]
      properties:
        workspace_id: {type: string}
        slug: {type: string}
        display_name: {type: string}
        description: {type: string}
        category: {type: string}
    AgentDraft:
      type: object
      required: [agent_id, revision, spec, validation_status, validation_findings, updated_by]
      properties:
        agent_id: {type: string}
        revision: {type: integer}
        spec: {type: object}
        validation_status: {type: string, enum: [valid, invalid]}
        validation_findings:
          type: array
          items: {$ref: "#/components/schemas/ValidationFinding"}
        updated_by: {type: string}
    UpdateDraftRequest:
      type: object
      required: [spec]
      properties:
        spec: {type: object}
    ValidationFinding:
      type: object
      required: [path, message]
      properties:
        path: {type: string}
        message: {type: string}
    AgentVersion:
      type: object
      required: [id, agent_id, version, source_draft_revision, spec, spec_digest]
      properties:
        id: {type: string}
        agent_id: {type: string}
        version: {type: integer}
        source_draft_revision: {type: integer}
        spec: {type: object}
        spec_digest: {type: string}
    AgentVersionList:
      type: object
      required: [items, page_info]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/AgentVersion"}}
        page_info: {$ref: "#/components/schemas/PageInfo"}
    SubmitReviewRequest:
      type: object
      required: [release_notes]
      properties:
        release_notes: {type: string, maxLength: 4000}
    ReviewDecisionRequest:
      type: object
      required: [decision, reason]
      properties:
        decision: {type: string, enum: [approve, reject]}
        reason: {type: string, maxLength: 4000}
    RollbackRequest:
      type: object
      required: [version_id]
      properties:
        version_id: {type: string}
    DiffEntry:
      type: object
      required: [path, change, category, risk]
      properties:
        path: {type: string}
        change: {type: string, enum: [added, removed, changed]}
        category: {type: string, enum: [metadata, behavior, security]}
        risk: {type: string, enum: [low, medium, high]}
        before: {}
        after: {}
    RiskSummary:
      type: object
      required: [total, high, medium, low]
      properties:
        total: {type: integer}
        high: {type: integer}
        medium: {type: integer}
        low: {type: integer}
    AgentReview:
      type: object
      required: [agent_id, draft_revision, draft_digest, release_notes, diff, risk_summary, status]
      properties:
        id: {type: string}
        agent_id: {type: string}
        draft_revision: {type: integer}
        draft_digest: {type: string}
        base_version_id: {type: string}
        base_version: {type: integer}
        release_notes: {type: string}
        diff: {type: array, items: {$ref: "#/components/schemas/DiffEntry"}}
        risk_summary: {$ref: "#/components/schemas/RiskSummary"}
        status: {type: string, enum: [not_submitted, pending, approved, rejected, superseded]}
        submitted_by: {type: string}
        reviewed_by: {type: string}
        review_reason: {type: string}
        submitted_at: {type: string, format: date-time}
        reviewed_at: {type: string, format: date-time}